Use transposition in all classes
Looks like individual accidentals (_, ^) for notes should not be used
when the key for the tune already has these accidentals at the start of
the line. This means halftones are not self-contained...

So if we ever make a parser for abc this has to be considered. Also for
output we should leave out accidentals. This involves storing the notes
and which accidentals are used for which key.