Fix rendering of tunes with accidentals
If we have a key with accidentals and individual tones with accidentals
identical to the key -- or tones *without* accidentals that match notes
*with* accidentals of the key, we need to either remove accidentals in
the generated output (to avoid accidentals at the key which are also at
the start of the line) or to add a natural accidental to the key which
doesn't match the accidentals of the tune.
This is implemented with an as_abc method of the Halftone which gets the
key as a parameter. The key now computes all accidentals dynamically.
Fix registration of keys so that keys are essentially a singleton.
Register voices with the tune so that we can navigate from the Tone to
tune.key.