Implement Cantus Firmus with tone-length !=1
Add iterator to checks: Harmony checks can now have several tones in the
Cantus Firmus in parallel to a single tone of the Contrapunctus.
Convert unit if the unit of the given --cantus-firmus is not 8.
Allow unnamed voice for Cantus Firmus if there is only one voice.
Also fix the abc parser to support unnamed voice.
For pre-checks of given --cantus-firmus we now allow to turn on
explanation via --explain-cp-cf option
We allow to ignore ugliness during this check and during depth first
search with --allow-ugliness.
Fix parsing of adjacent notes in a bar when reading abc file.
Fix a test that rejected a tune with a single voice without name for the
cantus firmus option.
Fix several tests, now each tone must have a bar.
Add rule in Makefile for generating .abc from .log.