Restore old eval computation
Now all except one long-running test produce the same result.
The fixes to the eval methods produce the same results for 8/8 notes in
the cantus firmus, and most of the examples are of that kind.

Add an option to set the print-frequency.
Print some results (notably the index of the best individual) in the
gene printing method.